requirements.txt
requests==2.31.0
requests-html==0.10.0
openai==1.12.0
python-dotenv==1.0.1
pytest==8.0.2
pytest-bdd==7.0.1
pytest-html==4.1.1
mlflow==2.10.2
plotly>=5.13.0
pandas>=1.5.0 